Improved efficency: calcBgrHsv.cpp::toHsvCV()

+++ examples/example4/calcBgrHsv.h
@@ -10,9 +10,11 @@ class calcBgrHsv{
       ~calcBgrHsv();
 
       double diffBGR(uchar b1, uchar g1, uchar r1, uchar b2, uchar g2, uchar r2);
-      double diffHSV(int h1, int s1, int v1, int h2, int s2, int v2);
-      void toHsvCV(int* hsv, uchar b, uchar g, uchar r);
+      double diffHSV(uchar h1, uchar s1, uchar v1, uchar h2, uchar s2, uchar v2);
+      void toHsvCV(uchar* hsv, uchar b, uchar g, uchar r);
       
 }; 
 #endif
+
+
